Update credits, readme, some docs, + move stuff
