+++ docs/EchoExtender.md
@@ -1,7 +1,9 @@
 # Echo Extender Usage Guide
 This guide will show off the process of creating an echo step-by-step and tell the user how to modify it to their will.
+
 ### Step 1 - Placing the Echo
 To place the echo, enter the game with your mod enabled, and navigate to the room you want to spawn the echo in. Place an `EEGhostSpot` placable object there. This object functions the same as a regular `GhostSpot` but it will not crash the game if the mod is disabled, unlike `GhostSpot`s.
+
 ### Step 2 - Writing Dialogue (Required)
 In order to give the echo dialogue, navigate to the region folder where you want the echo to spawn and place an `echoConv.txt` file right next to your `world_XX.txt` file. Be sure to store backups of your echo dialogue as this file will get encrypted, much like data pearls.
 
@@ -11,6 +13,7 @@ Translation files can be placed in text/text_\<lang>/echoConv_XX.txt, with \<lan
 Writing echo dialogue is essentially the same as writing writing lore pearl text. Dialogue is split into dialogue boxes - which are seperated by an empty line in the dialogue file - and these boxes can be split into two lines - which can be seperated using the `<LINE>` directive in the dialogue file.
 
 Dialogue boxes can be made conditional (making it only appear for a certain difficulty) by adding brackets in front of the dialogue line and adding the required difficulty for those dialogue to appear (ex.: `(Yellow)Hello World` will only appear for monk). For further details see [the example dialogue file.](echoConv.txt)
+
 ### Step 3 - Modifying Echo Settings
 SIDENOTE: As of 1.0 the settings file is necessary to specify room
