+++ src/Modules/Misc/_Module.cs
@@ -6,8 +6,6 @@ internal static class _Module
 
 	public static void Setup()
 	{
-		_Enums.Register();
-		ExtendedGates.InitExLocks();
 	}
 	public static void Enable()
 	{
@@ -16,13 +14,13 @@ internal static class _Module
 		SunBlockerFix.Apply();
 		GhostEffectColorsFix.Apply();
 		//CloudAdjustment.Apply();
-		ExtendedGates.Enable();
 		SuperstructureFusesHook.Apply();
 		MoreFadePalettes.Apply();
 		SlugcatRoomTemplates.Apply();
 		RainSong.Enable();
 		FadePaletteCombiner.Enable();
 		DecalPreview.Enable();
+		CustomSSSong.Enable();
 
 		On.RoomPalette.GetColor += RoomPalette_GetColor;
 	}
@@ -33,13 +31,13 @@ internal static class _Module
 		SunBlockerFix.Undo();
 		GhostEffectColorsFix.Undo();
 		//CloudAdjustment.Undo();
-		ExtendedGates.Disable();
 		SuperstructureFusesHook.Undo();
 		MoreFadePalettes.Undo();
 		SlugcatRoomTemplates.Undo();
 		RainSong.Disable();
 		FadePaletteCombiner.Disable();
 		DecalPreview.Disable();
+		CustomSSSong.Disable();
 	}
 
 	private static Color RoomPalette_GetColor(On.RoomPalette.orig_GetColor orig, ref RoomPalette self, RoomPalette.ColorName colorName)
