+++ src/Modules/Objects/BigWaterWheel.cs
@@ -1,6 +1,7 @@
 ﻿using System.Globalization;
 using System.IO;
 using System.Text.RegularExpressions;
+using RegionKit.API;
 using Watcher;
 
 namespace RegionKit.Modules.Objects
